Add JWT authentication and authorization

+++ Controllers/HotelsController.cs
@@ -1,12 +1,17 @@
 using System.ComponentModel.DataAnnotations;
+using System.Security.Claims;
+using System.Text.Json;
 using booking_api.Database;
 using booking_api.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace booking_api.Controllers;
 
 public record HotelWithoutId(string Name, string Location);
+
+public record RealmAccess(string[] roles);
 
 [Route("/hotels")]
 [Produces("application/json")]
@@ -33,9 +38,16 @@ public class HotelsController(AppDbContext db) : Controller
     }
 
     [HttpPost]
+    [Authorize]
     [ProducesResponseType(StatusCodes.Status201Created, Type = typeof(Hotel))]
     public async Task<IActionResult> PostHotel([FromBody] HotelWithoutId hotel)
     {
+        var roles = JsonSerializer.Deserialize<RealmAccess>(User.FindFirstValue("realm_access") ?? "{}")?.roles;
+        if (roles == null || !roles.Contains("create-hotels"))
+        {
+            return Unauthorized(new { Message = "You do not have the required roles" });
+        }
+
         var hotelId = Guid.NewGuid();
         var newHotel = new Hotel
         {
@@ -51,10 +63,17 @@ public class HotelsController(AppDbContext db) : Controller
     }
 
     [HttpDelete("{hotelId}")]
+    [Authorize]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> DeleteHotels(Guid hotelId)
     {
+        var roles = JsonSerializer.Deserialize<RealmAccess>(User.FindFirstValue("realm_access") ?? "{}")?.roles;
+        if (roles == null || !roles.Contains("delete-hotels"))
+        {
+            return Unauthorized(new { Message = "You do not have the required roles" });
+        }
+
         var rowsDeleted = await db.Hotels
             .Where(hotel => hotel.Id == hotelId)
             .ExecuteDeleteAsync();
@@ -74,10 +93,3 @@ public class HotelsController(AppDbContext db) : Controller
         return await hotel.ToListAsync();
     }
 }
-
-
-
-
-
-
-
